Lock an ICB slot actually by meram_lock_icb() and add meram_lock_trylock_icb()
This changes the behavior of meram_lock_icb() for sharing MERAM
between applications. meram_lock_icb() should acquire an exclusive
lock for a specified ICB slot and wait until available if it is
busy.
This also adds a new API 'meram_trylock_icb' to lock an ICB slot
only if it is free at the time of invocation. The new api may
be useful to look for an unused ICB.

Unfortunately this is preliminary and the exclusive lock
acquired by meram_lock_icb() must have effect only in a process.